requirements.txt
Django==2.2.17
Pillow==7.1.0
mysqlclient==1.4.2
django-ckeditor==5.6.1
django-naomi==0.8
django-auditlog==0.4.7
django-parler==2.2
python-decouple==3.3
